Add a README file and support for Jinja2 templates
Added a README file for the viewers(if any ..) of this repository to
help them understand the purpose of this software.
Added 2 Jinja2 templates in the app. One is the the generic base
template that most of the pages in the app will inherit from and another
is for the view function "index".
Both the templates are dynamic in nature due to the control structures in them
that helps to load pages differently for different data that is provided to templates.