Add a configuration file and login form for the app
Added a configuration file for the separation of concerns. All the
configuration values will be read from the configuration file.
Configuration values from environment variables are prefferd, if not
defined then hardcoded values are used as a fallback.

Added a login form that logs the users in and displays a message on the
screen to indicate the process. It only shows a message as users are not
yet supported for the app.

Used the flash() and get_flashed_messages() function to produce the
error messages.

Used the url_for() function to generate urls as it it much more reliable
than the hardcoded urls.

This thing probably shouldn't be here(Commit messages are not for this
dumb stuff), but I don't think anyone will ever read this.
Will this ever be read? Will this project be any good?
Are we doing it too early? Are we good enough?
Should we continue or leave it?
We will try our best anyways.

And we haven't thought about any name for the project yet.