Add database support for users and a requirements file
Add a module DBcm which allows us to hook into "with" statement
and make working with databases very efficient.
Add a models class for users which load the user information from the
database and write to their objects which make working with users a lot
easier.
The database connection details are loaded from environment variables.
Add a requirements file as this project depends upon a lot of third
party libraries.