Integrate flask_login and add a Posts model class
Integrated flask-login with the application which keeps record of the
signed in users and also implement the remember me functionality that
allows users to remain signed in even after closing their browsers.
Applied methods in the users class to store hashes of password instead
of the password itself.
Added a Posts model class for the "posts" table in the database.
Updated the requirements.txt to include flask-login.