Add a registration form for new users
Created a registration form for new users to register with the
application.
Still need to write the code to add newly registered users to the database.
Created the template for registration which provides the registration
form.
Added a view function to support registration.
Modified the login template to show a link for new registrations.