Improve the user profile page with user avatar, aboutme , last seen
Add support for much better profile pages with user avatar from
gravatar and displaying info such as about me and last seen.
Implemented a method "update" in the Users class that will update the
database with the latest details of the user when invoked.
Add support for about_me and last_seen in the database.
Update the "get_user" function to support the fields aboutme and last
seen.
Used a Jinja2 sub_template for the first time.