Add custom error pages and support for emails to admins
Added error handlers in the application and custom error pages that show
up whenever something goes wrong.
Added username validation in the edit profile form that gives an error
when a user tries to change his name to a name that is already in use by
some other user.
Added support for logging to files and sending email messages to admins
whenever an error happens in the application.