Add support for writing posts and to find users to follow.
Add a form for users to write posts on the website.
Added a explore page to the website where users can see all the posts
that are there and the usernames act as a link to the profile of the
users, from where we can follow each other.
Implemented a new function get_posts which returns all the posts
arranged according to the timestamp of the posts.