Add support for showing local time to users instead of utc
Added support for flask-moment in the application which shows timestamp
to the users according to his local time instead of the utc-timestamps
that are stored on the server/database.

Updated "requirements.txt" file to include all the libraries required for
the app to run.

Updated the templates to work with flask-moment.