Add a page for sharing assignments
Added two view functions that will help in sharing of assignments.
These functions take the subject name and read a csv file with the name
of subject and then read the data from it which has links to assignments
stored on cloud.
Then a page is rendered with these links.